+++ src/icv_trace/__init__.py
@@ -42,4 +42,4 @@ __all__ = [
     "trace",
 ]
 
-__version__ = "0.1.0rc2"
+__version__ = "0.1.0rc3"